~ reorganized context initialization so opengl is ready ~ moved pixmap, gc and image creation to the initialization   TODO: SUPPORT SCREEN SIZE CHANGES (THIS MIGHT NEED REINITIALIZATION OF THINGS) ~ moved fbo creation to initialization as it won't change anymore ~ reverted render code to the original loop as now the framebuffer is exactly as big as the whole display ~ moved glReadPixels off the wallpaper as that's code is exclusively used for screen rendering and not general rendering ~ XChangeProperty forces the update of the background, otherwise the compositor stops refreshing the screen and the background gets stuck ~ updated viewport variables to be integers instead of floats ~ reverted frame rendering code to be simpler and added the option to specify a framebuffer as target instead of screen

+++ src/WallpaperEngine/Render/CScene.cpp
@@ -56,7 +56,7 @@ CCamera* CScene::getCamera () const
     return this->m_camera;
 }
 
-void CScene::renderFrame (glm::vec4 viewport)
+void CScene::renderFrame (glm::ivec4 viewport)
 {
     auto projection = this->getScene ()->getOrthogonalProjection ();
     auto cur = this->m_objects.begin ();
@@ -82,7 +82,7 @@ void CScene::renderFrame (glm::vec4 viewport)
     glViewport (0, 0, projection->getWidth (), projection->getHeight ());
 }
 
-void CScene::updateMouse (glm::vec4 viewport)
+void CScene::updateMouse (glm::ivec4 viewport)
 {
     // projection also affects the mouse position
     auto projection = this->getScene ()->getOrthogonalProjection ();
